Delete vsproject/x64/Release/BoingBallSaver_v1.2.scr.zip
I found no issue with this file except it got corrupted and I thought I would re-upload with a different name. IOne user reported it as infected but I found no such infections scanning with Windows 11 Defender. However, I could not extract the archive. It errored out with "Invalid Archive" so a reupload it is. Slightly different naming too.